Combine parameters functions in various modules
